refactor: move shared functions to separate module to fix circular imports
