+++ discord_bot.py
@@ -4,7 +4,7 @@ import os
 from datetime import datetime
 import pytz
 import re
-from main import get_semester_collections, get_system_config
+from shared_utils import get_semester_collections, get_system_config
 import custom_print
 
 class LessonBot(commands.Bot):
